Add infrastructure to write unit tests.
It's very prelimnary and rudimentary. Have to know the init order of
the sub-systems and which variables to set etc. to make it work.

Ideally we should be able to divide the tests into these categories:

 1. In memory only e.g., locking, buffer pool, mutexes etc.
 2. IO related tests
 3. Mix of 1 & 2.

The locking uses the buffer pool as a backing store so we need to
startup the buffer pool even for trivial tests. It will be better to
make the locking memory allocator customizable so that we don't have
to initialize the buffer pool to test the locking.

This will be done as the next step.